+++ imports/ui/partials/Header.html
@@ -14,9 +14,10 @@
       	<li class="{{isActiveRoute 'memo.home'}}">
           <a class="{{isActiveRoute 'memo.home'}}" href="{{pathFor route='memo.home'}}">Memos</a>
         </li>
-        <li class="{{isActiveRoute 'memo.archive'}}">
-          <a href="{{pathFor route='memo.archive'}}" class="waves-effect {{isActiveRoute 'memo.archive'}}">
-            Archive
+        <li class="{{isActiveRoute 'memo.notifications'}}">
+          <a href="{{pathFor route='memo.notifications'}}" class="waves-effect {{isActiveRoute 'memo.notifications'}}">
+            <i class="fa fa-bell"></i>
+            Notification
         {{#if notificationCount}}
           <span class="chip blue lighten-4">{{notificationCount}}</span>
         {{/if}}
@@ -42,10 +43,10 @@
                 Memos
               </a>
             </li>
-            <li class="{{isActiveRoute 'memo.archive'}}">
-              <a href="{{pathFor route='memo.archive'}}" class="waves-effect {{isActiveRoute 'memo.archive'}}">
-                <i class="fa fa-archive"></i>
-                Archive
+            <li class="{{isActiveRoute 'memo.notifications'}}">
+              <a href="{{pathFor route='memo.notifications'}}" class="waves-effect {{isActiveRoute 'memo.notifications'}}">
+                <i class="fa fa-bell"></i>
+                Notification
             {{#if notificationCount}}
               <span class="chip blue lighten-4">{{notificationCount}}</span>
             {{/if}}
