+++ imports/ui/partials/Memo.html
@@ -3,7 +3,7 @@
 		<div class="card hoverable">
 		 		<div class="memo-tooltip center truncate {{#if isHovered}}tooltip-show{{/if}}">{{url}}</div>
         	  <div class="card-image center">
-              <i class="fa fa-archive expired-memo-icon {{#if isMemoExpired}}show{{else}}hide{{/if}}-expired"></i>
+              <i class="fa fa-clock-o expired-memo-icon {{#if isMemoExpired}}show{{else}}hide{{/if}}-expired"></i>
             <div class="expire-progress {{#if shouldExpireProgressbarShow}}show{{else}}hide{{/if}}-progress-bar">
               <div class="progress-bar progress-{{progressBarColor}}" style="width:{{expireStatus}}%"></div>
             </div>
@@ -40,10 +40,6 @@
               {{#afModal title=name collection="Memos" operation="update" doc=_id fields="labelId" formId="updateMemo"}}
                 <i class="fa fa-cog card-option"></i>
               {{/afModal}}
-              {{#if isMemoExpired}}
-              {{else}}
-              <i class="fa fa-archive archive-memo card-option"></i>
-              {{/if}}
               <i class="fa heart {{#if shouldFavoriteHightlight}}fa-heart{{ else }}fa-heart-o{{/if}}  card-option"></i>
               <i class="fa fa-share-square card-option" aria-hidden="true"></i>
           </div>
