+++ imports/ui/partials/Memo.js
@@ -113,9 +113,6 @@ TemplateController('Memo',{
 		},
 		'click .fa-share-square'(){
 			window.open(`https://twitter.com/intent/tweet?text=From my memo "${this.data.name}"&url=${this.data.url}`);
-		},
-		'click .archive-memo'(){
-			Meteor.call('archiveMemo', this.data);
 		}
 	},
 });
